#104  refactor: make videomenu into component
