+++ src/VideoComponents/css/MenuButton.css
@@ -20,7 +20,7 @@
 
 .transparentBtn {
   left: 10px;
-  background-size: 13px;
+  background-size: 15px;
 }
 
 .closeBtn {
