+++ src/VideoComponents/css/VideoApp.css
@@ -7,12 +7,3 @@
   height: 100%;
   border: solid 0.3px white;
 }
-
-.videoMenu {
-  height: 30px;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-  -webkit-user-select: none;
-  -webkit-app-region: drag;
-  position: relative;
-}
